+++ kernel/transaction/type/base/seata-at/src/main/java/org/apache/shardingsphere/transaction/base/seata/at/SeataATShardingSphereTransactionManager.java
@@ -30,7 +30,6 @@ import io.seata.tm.api.GlobalTransaction;
 import io.seata.tm.api.GlobalTransactionContext;
 import lombok.SneakyThrows;
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
-import org.apache.shardingsphere.transaction.core.ResourceDataSource;
 import org.apache.shardingsphere.transaction.api.TransactionType;
 import org.apache.shardingsphere.transaction.spi.ShardingSphereTransactionManager;
 
@@ -64,10 +63,10 @@ public final class SeataATShardingSphereTransactionManager implements ShardingSp
     }
     
     @Override
-    public void init(final Map<String, DatabaseType> databaseTypes, final Map<String, ResourceDataSource> resourceDataSources, final String providerType) {
+    public void init(final Map<String, DatabaseType> databaseTypes, final Map<String, DataSource> dataSources, final String providerType) {
         if (enableSeataAT) {
             initSeataRPCClient();
-            resourceDataSources.forEach((key, value) -> dataSourceMap.put(value.getOriginalName(), new DataSourceProxy(value.getDataSource())));
+            dataSources.forEach((key, value) -> dataSourceMap.put(key, new DataSourceProxy(value)));
         }
     }
     
@@ -84,13 +83,13 @@ public final class SeataATShardingSphereTransactionManager implements ShardingSp
     
     @Override
     public boolean isInTransaction() {
-        Preconditions.checkState(enableSeataAT, "sharding seata-at transaction has been disabled.");
+        Preconditions.checkState(enableSeataAT, "ShardingSphere seata-at transaction has been disabled.");
         return null != RootContext.getXID();
     }
     
     @Override
     public Connection getConnection(final String databaseName, final String dataSourceName) throws SQLException {
-        Preconditions.checkState(enableSeataAT, "sharding seata-at transaction has been disabled.");
+        Preconditions.checkState(enableSeataAT, "ShardingSphere seata-at transaction has been disabled.");
         return dataSourceMap.get(databaseName + "." + dataSourceName).getConnection();
     }
     
@@ -105,7 +104,7 @@ public final class SeataATShardingSphereTransactionManager implements ShardingSp
         if (timeout < 0) {
             throw new TransactionException("timeout should more than 0s");
         }
-        Preconditions.checkState(enableSeataAT, "sharding seata-at transaction has been disabled.");
+        Preconditions.checkState(enableSeataAT, "ShardingSphere seata-at transaction has been disabled.");
         GlobalTransaction globalTransaction = GlobalTransactionContext.getCurrentOrCreate();
         globalTransaction.begin(timeout * 1000);
         SeataTransactionHolder.set(globalTransaction);
@@ -114,7 +113,7 @@ public final class SeataATShardingSphereTransactionManager implements ShardingSp
     @Override
     @SneakyThrows(TransactionException.class)
     public void commit(final boolean rollbackOnly) {
-        Preconditions.checkState(enableSeataAT, "sharding seata-at transaction has been disabled.");
+        Preconditions.checkState(enableSeataAT, "ShardingSphere seata-at transaction has been disabled.");
         try {
             SeataTransactionHolder.get().commit();
         } finally {
@@ -126,7 +125,7 @@ public final class SeataATShardingSphereTransactionManager implements ShardingSp
     @Override
     @SneakyThrows(TransactionException.class)
     public void rollback() {
-        Preconditions.checkState(enableSeataAT, "sharding seata-at transaction has been disabled.");
+        Preconditions.checkState(enableSeataAT, "ShardingSphere seata-at transaction has been disabled.");
         try {
             SeataTransactionHolder.get().rollback();
         } finally {
